+++ client/ui/src/main/java/fr/esgi/pa/alliodesk/ui/controller/ErrorController.java
@@ -1,6 +1,6 @@
 package fr.esgi.pa.alliodesk.ui.controller;
 
-import fr.esgi.pa.alliodesk.core.Service;
+import fr.esgi.pa.alliodesk.core.models.Service;
 import fr.esgi.pa.alliodesk.core.request.AllioErrorRequest;
 import fr.esgi.pa.alliodesk.core.request.ServiceErrorRequest;
 import fr.esgi.pa.alliodesk.core.request.ServiceRequest;
